Moving content of __init__ to app.py and other minor adjustments.
