+++ timesketch/lib/experimental/win_services.py
@@ -109,6 +109,6 @@ def win_services(sketch_id):
 
 
 if __name__ == "__main__":
-    from timesketch import create_app
+    from timesketch.app import create_app
     with create_app().app_context():
         main()
