+++ timesketch/lib/tasks.py
@@ -33,7 +33,7 @@ from elasticsearch.exceptions import RequestError
 # Disabled until the project can provide a non-ES native import.
 # from mans_to_es import MansToEs
 
-from timesketch import create_celery_app
+from timesketch.app import create_celery_app
 from timesketch.lib import errors
 from timesketch.lib.analyzers import manager
 from timesketch.lib.datastores.elastic import ElasticsearchDataStore
