+++ timesketch/tsctl.py
@@ -35,7 +35,7 @@ from flask_script import prompt_pass
 from sqlalchemy.exc import IntegrityError
 from werkzeug.exceptions import Forbidden
 
-from timesketch import create_app
+from timesketch.app import create_app
 from timesketch.lib.datastores.elastic import ElasticsearchDataStore
 from timesketch.models import db_session
 from timesketch.models import drop_all
